requirements.txt
Flask==1.1.2
PyMySQL==0.9.3
google-cloud-ndb
googleapis_common_protos